delay demo modal on mousemove
